decimals from uint256 to uint8

+++ contracts/FujiOracle.sol
@@ -31,9 +31,9 @@ contract FujiOracle is IFujiOracle, Ownable {
   function getPriceOf(
     address _collateralAsset,
     address _borrowAsset,
-    uint256 _decimals
+    uint8 _decimals
   ) external view override returns (uint256 price) {
-    price = 10**_decimals;
+    price = 10**uint256(_decimals);
 
     if (_borrowAsset != address(0)) {
       price = price * _getUSDPrice(_borrowAsset);
